Add Pepecoin network
Test cases validated by importing the wif generated from the mnemonic into
Pepecoin core and checking the generated address against the address generated
by this code. By extension (no pun intended), the private and public keys must
also be derived accurately.

Note that Pepecoin, as a fork of Dogecoin, uses the same byte value for its WIF
prefix. This is unfortunate in some ways, but it also makes a good test of
other features in this distribution.